게시물 등록 & 평점 등록/조회 구현

+++ shopperman/src/main/java/com/example/shopperman/entity/Member.java
@@ -9,6 +9,6 @@ public class Member {
     private String name;
     private String nickname;
     private String addr;
-    private Double sumScore;      // 평점 총점
-	private Integer numScore;     // 평점 갯수
+    private Integer sumScore;      // 평점 총점
+	private Integer numScore;      // 평점 갯수
 }
